+++ src/Components/Fab/Fab.css
@@ -1,0 +1,24 @@
+.kc_fab_main_btn {
+  background-color: #f44336;
+  width: 60px;
+  height: 60px;
+  border-radius: 100%;
+  background: #f44336;
+  border: none;
+  outline: none;
+  color: #fff;
+  font-size: 36px;
+  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  transition: 0.3s;
+  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+  position: absolute;
+  bottom: 15px;
+  right: 15px;
+}
+
+.kc_fab_main_btn:focus {
+  transform: scale(1.1);
+  transform: rotate(45deg);
+  -ms-transform: rotate(45deg);
+  -webkit-transform: rotate(45deg);
+}
