+++ src/Components/Fab/Fab.js
@@ -1,0 +1,9 @@
+import './Fab.css'
+
+const Fab = () => {
+  return(
+    <button className="kc_fab_main_btn">+</button>
+  )
+}
+
+export default Fab;
