--- src/Components/ReadData/ReadData.js
@@ -1,1 +1,0 @@
-/* put the google sheet stuff from pyramid here */
